requirements.txt
# 核心依赖
redis>=5.0.1
kafka-python>=2.0.2
etcd3>=0.12.0
websockets>=12.0
pyyaml>=6.0.1
pytest>=7.4.0
typing-extensions>=4.8.0